Back End Services
Created the Maven project for Backend Services, configured to run on Port 8080.
Started the modelling of the Reservation and implementing Reservation controllers. Reservation exposes a post and get access point at "/post/reservations" and "get/reservations", this will return a list of Reservations via HTTP request. Post request will process the request for a new reservation, checking against reservations already made for that time of day to not overbook.
User, Role and RoleName classes setup but not implemented, to be implemented with authorisation.